backend/requirements.txt
fastapi==0.114.1
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
asyncpg==0.29.0
pydantic==2.9.2
python-dotenv==1.0.1
google-genai==0.3.0
Pillow==10.4.0
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
PyJWT[crypto]==2.9.0
httpx==0.27.2
google-auth==2.35.0
cryptography==43.0.1
